Qt 6.9+ forward-compatibility: explicit casts in QString::arg()/QChar()
Independent of the panelizer/Gerber feature. develop currently caps
the build at Qt 6.5.10; part of that cap is these source sites, which
stop compiling on Qt 6.9+ because the implicit enum/char16_t -> int
conversions in QString::arg() and the QChar(int) path were removed
upstream.

Each change is a minimal, behaviour-preserving explicit cast that is
equally valid on Qt 6.5:
- QString::arg(enum)            -> arg(static_cast<int>(enum))
    commands.cpp (Qt::Orientations), connectoritem.cpp / itembase.cpp /
    sketchwidget.cpp (ViewGeometry::WireFlags)
- QChar(enum)                   -> QChar(static_cast<int>(enum))
    paletteitembase.cpp (ViewLayer::ViewLayerPlacement)
- QString::arg(char16_t, ...)   -> arg(static_cast<int>(c), ...)
    resistor.cpp / textutils.cpp (MicroSymbolCode, QChar::unicode())

Offered separately so the maintainer can decide whether to take them
toward lifting the Qt version cap; they are not required by the
feature. Verified: feature + these casts build and link on Qt 6.11.1.

+++ src/items/itembase.cpp
@@ -2006,7 +2006,7 @@ void ItemBase::debugInfo2(const QString & msg) const
 	                   .arg(this->instanceTitle())
 	                   .arg(this->viewLayerID())
 	                   .arg(this->viewLayerPlacement())
-	                   .arg(this->wireFlags())
+	                   .arg(static_cast<int>(this->wireFlags()))
 			   .arg((long) dynamic_cast<const QGraphicsItem *>(this), 0, 16)
 	                   .arg(m_viewID)
 	                   .arg(this->zValue())
